chore: refactor v2/v3 codes (see desc)
- renamed privacy to obfuscate
- split signature.ts in both v2/v3 into verify.ts and wrap.ts
- /privacy/salt was only used in v2, so moved over to v2
- added validate into shared folder
- renamed v3's validate to validateW3C and placed it in shared/validate, as it only validate specific fields for w3c compliance
