--- src/shared/privacy/index.ts
@@ -1,1 +1,0 @@
-export * from "./privacy";
