+++ src/v2/obfuscate.test.ts
@@ -1,6 +1,6 @@
-import { obfuscateData, obfuscateDocument } from "./privacy";
-import { WrappedDocument, SchemaId } from "../../shared/@types/document";
-import { getData } from "../utils";
+import { obfuscateData, obfuscateDocument } from "./obfuscate";
+import { WrappedDocument, SchemaId } from "../shared/@types/document";
+import { getData } from "../shared/utils";
 
 describe("privacy", () => {
   describe("obfuscateData", () => {
