--- src/v2/signature/index.ts
@@ -1,1 +1,0 @@
-export * from "./signature";
